+++ src/loop-analysis/loop-state.hpp
@@ -29,6 +29,7 @@
 
 #include "mapping/loop.hpp"
 #include "workload/problem-config.hpp"
+#include "workload/operation-space.hpp"
 
 namespace analysis
 {
