+++ src/problem-shape/per-data-space.hpp
@@ -27,7 +27,11 @@
 
 #pragma once
 
+#include <boost/archive/text_oarchive.hpp>
+#include <boost/archive/text_iarchive.hpp>
+
 #include "util/dynamic-array.hpp"
+#include "global-names.hpp"
 
 namespace problem
 {
