+++ src/problem-shape/problem-config.cpp
@@ -27,40 +27,34 @@
 
 #include <algorithm>
 
-#include "cnn/problem-config.hpp"
+#include "loop-analysis/point-set.hpp"
+
+#include "problem-config.hpp"
+#include "workload-config.hpp"
+#include "operation-space.hpp"
 
 namespace problem
 {
 
+// ======================================== //
+//              Problem Shape               //
+// ======================================== //
+
+// Globals
 unsigned NumDimensions;
+std::map<Dimension, std::string> DimensionName;
+std::map<char, Dimension> DimensionID;
+
 unsigned NumDataSpaces;
 std::map<DataSpaceID, std::string> DataSpaceIDToName;
 std::map<std::string, DataSpaceID> DataSpaceNameToID;
 std::vector<unsigned> DataSpaceOrder;
 
 std::function<bool(const DataSpaceID d)> IsReadWriteDataSpace;
-std::vector<std::function<Point(WorkloadConfig*, const OperationPoint&)>> projectors;
 
-// std::ostream& operator << (std::ostream& out, const DataSpaceID& d)
-// {
-//   out << DataSpaceIDToName[d];
-//   return out;
-// }
+std::vector<std::function<Point(WorkloadConfig*, const OperationPoint&)>> Projectors;
 
-std::map<Dimension, std::string> DimensionName;
-std::map<char, Dimension> DimensionID;
-
-// std::ostream& operator << (std::ostream& out, const Dimension& dim)
-// {
-//   out << DimensionName[dim];
-//   return out;
-// }
-
-// ======================================== //
-//              Problem Shape               //
-// ======================================== //
-
-void BuildProblemShape()
+void ParseProblemShape()
 {
   NumDimensions = 7;
   NumDataSpaces = 3;
@@ -87,14 +81,6 @@ void BuildProblemShape()
     Num
   };
 
-  // enum class DataSpaceID : unsigned int
-  // {
-  //   Weight,
-  //   Input,
-  //   Output,
-  //   Num
-  // };
-  
   DataSpaceIDToName = {
     {0, "Weights"},
     {1, "Inputs"},
@@ -134,7 +120,7 @@ void BuildProblemShape()
                  {'K', 5 },
                  {'N', 6 }, };
 
-  projectors =
+  Projectors =
     {
       [](WorkloadConfig* wc, const OperationPoint& problem_point)
       {
@@ -181,112 +167,6 @@ void BuildProblemShape()
       }
     };  
 }
-
-// ======================================== //
-//             OperationSpace               //
-// ======================================== //
-
-OperationSpace::OperationSpace(WorkloadConfig* wc) :
-    workload_config_(wc)
-{
-  for (unsigned space_id = 0; space_id < NumDataSpaces; space_id++)
-    data_spaces_.push_back(DataSpace(DataSpaceOrder.at(space_id)));
-}
-
-OperationSpace::OperationSpace() :
-    OperationSpace(nullptr)
-{ }
-
-OperationSpace::OperationSpace(WorkloadConfig* wc, const OperationPoint& low, const OperationPoint& high) :
-    workload_config_(wc)
-{
-  for (unsigned space_id = 0; space_id < NumDataSpaces; space_id++)
-  {
-    auto space_low = projectors.at(space_id)(workload_config_, low);
-    auto space_high = projectors.at(space_id)(workload_config_, high);
-    // Increment the high points by 1 because the AAHR constructor wants
-    // an exclusive max point.
-    space_high.IncrementAllDimensions();
-    data_spaces_.push_back(DataSpace(DataSpaceOrder.at(space_id), space_low, space_high));
-  }
-}
-
-void OperationSpace::Reset()
-{
-  for (auto& d : data_spaces_)
-    d.Reset();
-}
-
-OperationSpace& OperationSpace::operator += (const OperationSpace& s)
-{
-  for (unsigned i = 0; i < data_spaces_.size(); i++)
-    data_spaces_.at(i) += s.data_spaces_.at(i);
-
-  return (*this);
-}
-
-OperationSpace& OperationSpace::operator += (const OperationPoint& p)
-{
-  for (unsigned i = 0; i < data_spaces_.size(); i++)
-    data_spaces_.at(i) += projectors.at(i)(workload_config_, p);
-
-  return (*this);
-}
-
-OperationSpace OperationSpace::operator - (const OperationSpace& p)
-{
-  OperationSpace retval(workload_config_);
-
-  for (unsigned i = 0; i < data_spaces_.size(); i++)
-    retval.data_spaces_.at(i) = data_spaces_.at(i) - p.data_spaces_.at(i);
-  
-  return retval;
-}
-
-PerDataSpace<std::size_t> OperationSpace::GetSizes() const
-{
-  PerDataSpace<std::size_t> retval;
-  
-  for (unsigned i = 0; i < data_spaces_.size(); i++)
-    retval.at(i) = data_spaces_.at(i).size();
-
-  return retval;
-}
-
-std::size_t OperationSpace::GetSize(const int t) const
-{
-  return data_spaces_.at(t).size();
-}
-
-bool OperationSpace::IsEmpty(const int t) const
-{
-  return data_spaces_.at(t).empty();
-}
-
-bool OperationSpace::CheckEquality(const OperationSpace& rhs, const int t) const
-{
-  return data_spaces_.at(t) == rhs.data_spaces_.at(t);
-}
-
-void OperationSpace::PrintSizes()
-{
-  for (unsigned i = 0; i < data_spaces_.size()-1; i++)
-    std::cout << DataSpaceIDToName[i] << " = " << data_spaces_.at(i).size() << ", ";
-  std::cout << DataSpaceIDToName[data_spaces_.size()-1] << " = " << data_spaces_.back().size() << std::endl;
-}
-
-void OperationSpace::Print() const
-{
-  for (auto& d : data_spaces_)
-    d.Print();
-}
-
-void OperationSpace::Print(DataSpaceID pv) const
-{
-  auto& d = data_spaces_.at(unsigned(pv));
-  d.Print();
-}
-
 
 PerDataSpace<std::size_t> GetMaxWorkingSetSizes(
     problem::PerProblemDimension<int> dimension_sizes)
